testing gem config

+++ container/src/tools/move_file.rb
@@ -8,6 +8,10 @@ module Tools
     param :destination, desc: "The relative path to move the file or directory to."
 
     def execute(source:, destination:)
+      if ENV["MILDRED_DRY_RUN"] == "1"
+        return "[DRY RUN] Would move #{source} to #{destination}"
+      end
+
       FileUtils.mv(source, destination)
       "Moved #{source} to #{destination}"
     rescue => e
